Update log.md
Create a new branch called Setting.
This branch is used for developing setting menu related function.
For instance, windows size,preference,storage position,or anything not belongs to main function.
This branch is a prototype. That means if main structure is completed, this branch will come to its end. Any further function should be implemented in next more focused branch.