Floor mover
When floor is up to the max-floor instead of creating the over-max floor it would overwrite below floor by current floor content. However, due to the nature of cloning is slow and violating Content-Security-Policy leads to a bombing of error messages. I will immediately rewrite fsfloor function from fixed to dynamic method