Fix CSP and performance issue
Rewrite fsfloor sign function with remove and append element. It's not the fastest way to achieve this. But god-damned electron not allow me to do it by my way due to some 'in-line style writing violate bruh bruh'.
Anyway, it could be improved by better remove and append function. By far, this is it.