Build a demo function mntgroupload for testing monitored groups loading
I make a function called mntgroupload. This function will handle all user-created monitored groups loading. When clicking header, it would automatically detect if the content is loaded, if not, load it or leave it.
This function not been fully tested because of  the lack of members in self-created monitored groups.
I will fully test it after the custom movemenu is complete